added clientside + server fuctions
added functions to index.js get the user records and add a new user record to database.

added clientside function newUser which takes a username and checks if it already exists in the database.

If it doesnt it will call function createnewUser to add the new user to the database